feat: separate passkey key storage logic
