+++ index.html
@@ -9,11 +9,8 @@
 </head>
 <body>
     <h1>YouTube Video Summarizer</h1>
-    <p>Enter a YouTube URL and (on first use) your Together API key to summarize the video transcript. The key is stored securely in your browser and can be reset at any time.</p>
+    <p>Enter a YouTube URL to summarize its transcript. Configure your Together API key on the <a href="settings.html">settings page</a>.</p>
     <input type="text" id="url" placeholder="YouTube URL">
-    <input type="text" id="apiKey" class="api" placeholder="Together API Key">
-    <button id="decryptKey" class="api" style="display:none;">Decrypt with passkey</button>
-    <button id="resetKey" class="api" style="display:none;">Reset API Key</button>
     <button id="summarize">Summarize</button>
     <div id="status"></div>
     <h2>Summary</h2>
